big speed improvements
