+++ vasprun.py
@@ -12,152 +12,10 @@ import pickle
 
 
 
-class Vasprun:
-
-    def __init__(self, path: Path):
-        self.path = path
-        self.root = read_vasprun(path)
 
 
-    @property
-    def eigenvalues(self) -> pd.DataFrame:
-        spins = self.root.find('calculation').find(
-            'eigenvalues').find('array').find('set').findall('set')
-        eigenvalues = pd.DataFrame()
-        for spin in spins:
-            spin_index = int(spin.attrib['comment'].split()[-1])
-
-            kpoints = spin.findall('set')
-            for kpoint in kpoints:
-                kpoint_index = int(kpoint.attrib['comment'].split()[-1])
-                rows = [val.text for val in kpoint.findall('r')]
-                energies = [float(row.split()[0]) for row in rows]
-                occupations = [float(row.split()[1]) for row in rows]
-                # create a dataframe for this kpoint
-                kpoint_df = pd.DataFrame(
-                    {'kpoint': kpoint_index, 'spin': spin_index, 'energy': energies, 'occupation': occupations})
-                # concatenate to the main dataframe
-                eigenvalues = pd.concat([eigenvalues, kpoint_df])
-
-        return eigenvalues
 
 
-    @property
-    def dos(self):
-        return DensityOfStates(self)
-
-    @property
-    def bands(self):
-        return BandStructure(self)
-
-    @property
-    def structure(self):
-        return Structure(self)
-
-
-class DensityOfStates:
-
-    def __init__(self, vasprun: Vasprun):
-
-        try:
-            self.dos_element = vasprun.root.find('calculation').find('dos')
-
-            self.total_dos_element = self.dos_element.find(
-                'total').find('array').find('set')
-            self.projected_dos_element = vasprun.root.find(
-                'calculation').find('projected')
-        except AttributeError:
-            print('No DOS data found in vasprun.xml')
-
-    @property
-    def total(self) -> pd.DataFrame:
-        total_dos = pd.DataFrame()
-        spins = self.total_dos_element.findall('set')
-        for spin in spins:
-            spin_index = int(spin.attrib['comment'].split()[-1])
-            vals = unpack_rarray(spin)
-            headers = ['energy', 'total', 'integrated']
-            df = pd.DataFrame(vals, columns=headers)
-            df['spin'] = spin_index
-            total_dos = pd.concat([total_dos, df])
-
-        return total_dos
-
-    @property
-    def partial(self) -> pd.DataFrame:
-        '''Returns a dataframe of the projected density of states.'''
-        partial_dos = pd.DataFrame()
-        total_dfs = []
-        partial_dos_element = self.dos_element.find('partial').find('array')
-        headers = [
-            field.text.strip() for field in partial_dos_element.findall('field')]
-        ions = partial_dos_element.find('set').findall('set')
-        for ion in ions:
-            ion_index = int(ion.attrib['comment'].split()[-1])
-
-            spins = ion.findall('set')
-            for spin in spins:
-                spin_index = int(spin.attrib['comment'].split()[-1])
-                vals = unpack_rarray(spin)
-                df = pd.DataFrame(vals, columns=headers)
-                df['ion'] = ion_index
-                df['spin'] = spin_index
-                total_dfs.append(df)
-
-        partial_dos = pd.concat([partial_dos, total_dfs])
-
-        return partial_dos
-
-    @property
-    def projected(self) -> pd.DataFrame:
-        '''Returns a dataframe of the projected density of states.'''
-
-        projected_dos_element = self.projected_dos_element.find('array')
-        headers = [field.text.strip()
-                   for field in projected_dos_element.findall('field')]
-        spins = [spin for spin in projected_dos_element.find(
-            'set').findall('set')]
-        n_spins = len(spins)
-        kpoints = [kpoint.findall('set') for kpoint in spins]
-        n_kpoints = len(kpoints[0])
-        band_lists = [band.findall('set')
-                      for kpoint in kpoints for band in kpoint]
-        band_arrays = [unpack_rarray(
-            band) for band_list in band_lists for band in band_list]
-        n_bands = len(band_lists[0])
-        n_ions = len(band_arrays[0])
-        ion_indices = np.arange(1, len(band_arrays[0]) + 1)
-        spin_indices = np.arange(1, n_spins + 1)
-        kpoint_indices = np.arange(1, n_kpoints + 1)
-        band_indices = np.arange(1, n_bands + 1)
-        dfs = [pd.DataFrame(band_array, columns=headers)
-               for band_array in band_arrays]
-        projected_dos = pd.concat(dfs)
-
-        projected_dos['ion'] = np.tile(
-            ion_indices, n_spins * n_kpoints * n_bands)
-        projected_dos['spin'] = np.tile(
-            np.repeat(spin_indices, n_kpoints * n_bands), n_ions)
-        projected_dos['kpoint'] = np.tile(
-            np.repeat(kpoint_indices, n_bands), n_ions * n_spins)
-        projected_dos['band'] = np.tile(
-            band_indices, n_ions * n_spins * n_kpoints)
-
-        return projected_dos
-
-    @property
-    def cprojected(self) -> pd.DataFrame:
-        ''' Checks to see if a vrproj.pkl file exists and if so, returns the projected density of states from that file. If not, it creates the file and returns the projected density of states.'''
-
-        if os.path.isfile('vrproj.pkl') and os.path.getsize('vrproj.pkl') > 0:
-            with open('vrproj.pkl', 'rb') as f:
-                projected_dos = pickle.load(f)
-        else:
-            projected_dos = self.projected
-            with open('vrproj.pkl', 'wb') as f:
-                pickle.dump(projected_dos, f)
-
-        return projected_dos
 
 
 class BandStructure:
